Add mobile navbar scaffold
